--- src/main/java/com/boardgo/domain/user/service/facade/UserQueryServiceFacadeInter.java
@@ -1,3 +1,0 @@
-package com.boardgo.domain.user.service.facade;
-
-public interface UserQueryServiceFacadeInter {}
